move fetch instant results thunk to search actions
https://coveord.atlassian.net/browse/KIT-1631
