+++ packages/headless/src/features/history/history-slice.test.ts
@@ -98,6 +98,7 @@ describe('history slice', () => {
         defaultFilters: {aq: '', cq: '', lq: '', dq: ''},
       },
       querySet: {foo: 'bar', hello: 'world'},
+      instantResults: {},
       sortCriteria: 'date descending',
       pipeline: 'my-pipeline',
       searchHub: 'my-search-hub',
