+++ packages/headless/src/features/instant-results/instant-results-actions.ts
@@ -1,21 +1,12 @@
-import {createAction, createAsyncThunk} from '@reduxjs/toolkit';
+import {createAction} from '@reduxjs/toolkit';
 import {InstantResultSection} from '../../state/state-sections';
 import {
   validatePayload,
   requiredEmptyAllowedString,
   requiredNonEmptyString,
 } from '../../utils/validate-payload';
-import {
-  isErrorResponse,
-  AsyncThunkSearchOptions,
-} from '../../api/search/search-api-client';
 import {StateNeededByExecuteSearch} from '../search/search-actions';
-import {mapSearchRequest, mapSearchResponse} from '../search/search-mappings';
-import {buildSearchAndFoldingLoadCollectionRequest} from '../search-and-folding/search-and-folding-request';
-import {logQueryError} from '../search/search-analytics-actions';
 import {Result} from '../../case-assist.index';
-import {NumberValue} from '@coveo/bueno';
-
 export interface RegisterInstantResultActionCreatorPayload {
   /**
    * The search box id.
@@ -76,56 +67,3 @@ export interface FetchInstantResultsThunkReturn {
 
 export type StateNeededByInstantResults = StateNeededByExecuteSearch &
   InstantResultSection;
-
-export const fetchInstantResults = createAsyncThunk<
-  FetchInstantResultsThunkReturn,
-  FetchInstantResultsActionCreatorPayload,
-  AsyncThunkSearchOptions<StateNeededByInstantResults>
->(
-  'instantResults/fetch',
-  async (
-    payload: FetchInstantResultsActionCreatorPayload,
-    {getState, dispatch, rejectWithValue, extra: {apiClient, validatePayload}}
-  ) => {
-    validatePayload(payload, {
-      id: requiredNonEmptyString,
-      q: requiredNonEmptyString,
-      maxResultsPerQuery: new NumberValue({
-        required: true,
-        min: 1,
-      }),
-    });
-    const q = payload.q;
-    const state = getState();
-
-    const {request, mappings} = await buildInstantResultSearchRequest(state, q);
-    const response = await mapSearchResponse(
-      await apiClient.search(request),
-      mappings
-    );
-
-    if (isErrorResponse(response)) {
-      dispatch(logQueryError(response.error));
-      return rejectWithValue(response.error);
-    }
-    return {
-      results: response.success.results.slice(
-        0,
-        payload.maxResultsPerQuery + 1
-      ),
-    };
-  }
-);
-
-export const buildInstantResultSearchRequest = async (
-  state: StateNeededByExecuteSearch,
-  q: string
-) => {
-  const sharedWithFoldingRequest =
-    await buildSearchAndFoldingLoadCollectionRequest(state);
-
-  return mapSearchRequest({
-    ...sharedWithFoldingRequest,
-    q,
-  });
-};
